requirements.txt
torch>=2.0.0
numpy
tqdm
pillow 
transformers